% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{four_corners_lcc_proj}
\alias{four_corners_lcc_proj}
\title{A nice projection for the Four Corners states}
\format{
A character string with the PROJ.4 projection information
}
\usage{
four_corners_lcc_proj
}
\description{
In PROJ.4:
"+proj=lcc +lat_1=41.0034439990657 +lat_2=32.9892379991201 +lat_0=37.0002329990929
+lon_0=-109.050431 +x_0=-109.050431 +x_y=-109.050431"
}
\keyword{datasets}
