% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{four_corners_states}
\alias{four_corners_states}
\title{The boundaries of the Four Corners states}
\format{
Simple feature collection with 4 features and 2 fields:
\describe{
  \item{STATEFP}{the state's FIPS code}
  \item{geometry}{the state's boundary}
 }
}
\usage{
four_corners_states
}
\description{
These data are derived from the US Census TIGRIS database.
}
\keyword{datasets}
