% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swus_hillshade_500m}
\alias{swus_hillshade_500m}
\title{A hillshade of the Four Corners states}
\format{
A data frame with 5,556,936 rows and 3 variables:
\describe{
  \item{x}{the easting of the hillshade value}
  \item{y}{the northing of the hillshade value}
  \item{ID}{the hillshade value}
 }
}
\usage{
swus_hillshade_500m
}
\description{
This hillshade is at 500m resolution, and in the four_corners_lcc_proj projection.
These data are appropriate for plotting with ggplot2.
}
\keyword{datasets}
