% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uusw_boundary}
\alias{uusw_boundary}
\title{The boundaries of the Upland US Southwest}
\format{Geometry set for 1 feature}
\usage{
uusw_boundary
}
\description{
The UUSW was geographically defined by Bocinsky et al. 2016 as the region between 105–113ºW and 32–38ºN.
}
\keyword{datasets}
