% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uusw_counties}
\alias{uusw_counties}
\title{The boundaries of the Upland US Southwest counties used in Robinson et al. 2020}
\format{
SSimple feature collection with 40 features and 2 field:
\describe{
  \item{County}{the county's name and state}
  \item{geometry}{the county's boundary}
 }
}
\usage{
uusw_counties
}
\description{
The UUSW was geographically defined by Bocinsky et al. 2016 as the region between 105–113ºW and 32–38ºN.
Robinson et al. 2020 refined that definition in two ways.
First, because location data for the radiocarbon database are only resolved to the county
level for most sites, the authors defined their study region to include all counties whose
geographic centroids fall within the region from Bocinsky et al. 2016.
Second, because of the generally poor tree-ring date record in the Sonoran Desert,
the authors roughly exclude the Phoenix and Tucson basins by dropping
Maricopa, Pinal, and Pima counties in Arizona.
}
\keyword{datasets}
